Enable private domains and rename file to `getPublicSuffix`
